+++ package.json
@@ -62,6 +62,7 @@
     "js-yaml-loader": "^1.2.2",
     "jss": "^10.9.0",
     "jss-preset-default": "^10.9.0",
+    "license-checker": "^25.0.1",
     "loader-utils": "^3.2.0",
     "localstorage-slim": "^2.2.0",
     "markdown-to-jsx": "^7.1.5",
