+++ src/components/AppRoot.ts
@@ -1,8 +1,8 @@
-import pkg from "../../package.json";
-import ons_pkg from "./../../node_modules/onsenui/package.json";
-import react_ons_pkg from "./../../node_modules/react-onsenui/package.json";
-import react_bs_pkg from "./../../node_modules/react-bootstrap/package.json";
-import react_pkg from "./../../node_modules/react/package.json";
+import pkg from "#Package";
+import ons_pkg from "#OnsenUI-Package";
+import react_ons_pkg from "#React-OnsenUI-Package";
+import react_bs_pkg from "#React-Bootstrap-Package";
+import react_pkg from "#React-Package";
 
 /**
  * Provides information about the app and frameworks
@@ -11,13 +11,6 @@ import react_pkg from "./../../node_modules/react/package.json";
 class AppRoot extends HTMLElement {
   public constructor() {
     super();
-    this.initConfigStats([
-      { key: "version", value: pkg.version },
-      { key: "onsenui-version", value: ons_pkg.version },
-      { key: "react-onsenui-version", value: react_ons_pkg.version },
-      { key: "react-version", value: react_pkg.version },
-      { key: "react-bootstrap-version", value: react_bs_pkg.version },
-    ]);
   }
 
   private initConfigStats(data: { key: string; value: string }[]) {
@@ -27,20 +20,13 @@ class AppRoot extends HTMLElement {
   }
 
   public connectedCallback() {
-    // Element wurde ins DOM eingehängt
-  }
-
-  public disconnectedCallback() {
-    // Element wurde entfernt
-  }
-
-  public adoptedCallback() {
-    // Element ist in ein anderes Dokument umgezogen
-  }
-
-  public attributeChangedCallback(name: any, oldValue: any, newValue: any) {
-    // Elementparameter wurden geändert
-    // Achtung attributeChangedCallback wird vor connectedCallback aufgerufen
+    this.initConfigStats([
+      { key: "version", value: pkg.version },
+      { key: "onsenui-version", value: ons_pkg.version },
+      { key: "react-onsenui-version", value: react_ons_pkg.version },
+      { key: "react-version", value: react_pkg.version },
+      { key: "react-bootstrap-version", value: react_bs_pkg.version },
+    ]);
   }
 }
